pyproject.toml
[project]
name = "thinksy"
version = "0.1.0"
description = "Multi-agent AI marketing pipeline for video ad generation and distribution"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "runwayml>=0.8.0",
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.2.0",
    "requests>=2.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "typer>=0.12.0",
    "rich>=13.9.0",
    "aiohttp>=3.10.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
thinksy = "pipeline.cli:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
